fix: do not display recipe details contents until loaded

+++ packages/frontend/src/app/pages/recipe-components/recipe/recipe.page.ts
@@ -52,7 +52,7 @@ export class RecipePage {
     release: () => void;
   } = null;
 
-  recipe: Recipe;
+  recipe: Recipe | null;
   recipeId: string;
   ingredients?: ParsedIngredient[];
   instructions?: ParsedInstruction[];
@@ -96,7 +96,7 @@ export class RecipePage {
       throw new Error("No recipeId was provided");
     }
     this.recipeId = recipeId;
-    this.recipe = {} as Recipe;
+    this.recipe = null;
 
     this.scale =
       this.recipeCompletionTrackerService.getRecipeScale(this.recipeId) || 1;
@@ -111,7 +111,7 @@ export class RecipePage {
   ionViewWillEnter() {
     const loading = this.loadingService.start();
 
-    this.recipe = {} as Recipe;
+    this.recipe = null;
 
     this.loadAll().then(
       () => {
@@ -200,6 +200,8 @@ export class RecipePage {
   }
 
   updateRatingVisual() {
+    if (!this.recipe) return;
+
     this.ratingVisual = new Array<string>(5)
       .fill("star", 0, this.recipe.rating)
       .fill("star-outline", this.recipe.rating, 5);
@@ -310,6 +312,8 @@ export class RecipePage {
   }
 
   applyScale() {
+    if (!this.recipe) return;
+
     this.ingredients = this.recipeService.parseIngredients(
       this.recipe.ingredients,
       this.scale,
@@ -355,6 +359,8 @@ export class RecipePage {
   }
 
   private async _deleteRecipe() {
+    if (!this.recipe) return;
+
     const loading = this.loadingService.start();
 
     const response = await this.recipeService.delete(this.recipe.id);
@@ -411,6 +417,8 @@ export class RecipePage {
   }
 
   async moveToFolder(folderName: RecipeFolderName) {
+    if (!this.recipe) return;
+
     const loading = this.loadingService.start();
 
     this.recipe.folder = folderName;
@@ -450,6 +458,8 @@ export class RecipePage {
   }
 
   async addLabel(title: string) {
+    if (!this.recipe) return;
+
     if (title.length === 0) {
       const message = await this.translate
         .get("pages.recipeDetails.enterLabelWarning")
@@ -509,6 +519,8 @@ export class RecipePage {
   }
 
   private async _deleteLabel(label: Label) {
+    if (!this.recipe) return;
+
     const loading = this.loadingService.start();
 
     await this.labelService.removeFromRecipe({
@@ -520,6 +532,8 @@ export class RecipePage {
   }
 
   async cloneRecipe() {
+    if (!this.recipe) return;
+
     const loading = this.loadingService.start();
     const response = await this.recipeService.create({
       ...this.recipe,
@@ -575,6 +589,8 @@ export class RecipePage {
   }
 
   async openImageViewer() {
+    if (!this.recipe) return;
+
     const imageViewerModal = await this.modalCtrl.create({
       component: ImageViewerComponent,
       componentProps: {
@@ -585,6 +601,8 @@ export class RecipePage {
   }
 
   pinRecipe() {
+    if (!this.recipe) return;
+
     this.cookingToolbarService.pinRecipe({
       id: this.recipe.id,
       title: this.recipe.title,
@@ -593,6 +611,8 @@ export class RecipePage {
   }
 
   unpinRecipe() {
+    if (!this.recipe) return;
+
     this.cookingToolbarService.unpinRecipe(this.recipe.id);
   }
